requirements.txt
flask
flask-cors
diffusers
transformers
accelerate
safetensors
huggingface-hub
pillow
numpy
